fix(benchmark): narrow non-orderkey identifiers to INTEGER in tpch_schema.sql
Spec clause 1.3.1 requires an identifier to hold every key value generated for
its column and support at least 2,147,483,647 unique values, and exempts
identifier columns from the datatype-consistency rule so each can be sized
individually. Orderkeys are sparsely populated across 6,000,000 * SF and pass
2^31 above ~SF358, so they stay BIGINT; partkey, suppkey and custkey are dense
and remain inside INTEGER well past SF1000.

This aligns the duckdb datasets with sirius-db/tpchgen-rs, which emits Int32
for every key except l_orderkey/o_orderkey, so parquet and duckdb share one key
layout -- and matches the dtypes the SF1000 compression plans were tuned for.
At SF1000 it removes 48 GB from lineitem's two key columns alone.

The tradeoff is that a database built here is no longer column-type
interchangeable with one built by DuckDB's tpch extension, which types all keys
BIGINT; the header comment now says so.

+++ test/tpch_performance/tpch_schema.sql
@@ -1,9 +1,21 @@
 -- TPC-H schema for loading classic dbgen .tbl output into DuckDB.
 --
--- Types match what DuckDB's tpch extension creates, so a database built from
--- dbgen and one built by the extension are interchangeable. That means keys are
--- BIGINT rather than dss.ddl's INTEGER (orderkeys pass 2^31 above ~SF300) and
--- fixed-width text is VARCHAR, which is how DuckDB stores CHAR anyway.
+-- Key widths follow spec clause 1.3.1: an identifier must hold every key value
+-- generated for its column and support at least 2,147,483,647 unique values.
+-- Orderkeys are sparsely populated across 6,000,000 * SF, so they exceed 2^31
+-- above ~SF358 and must be BIGINT. Every other key is densely populated and
+-- stays inside INTEGER well past SF1000 (partkey SF*200,000, custkey
+-- SF*150,000, suppkey SF*10,000). Clause 1.3.1 exempts identifier columns from
+-- the datatype-consistency rule so they can be sized individually.
+--
+-- This matches sirius-db/tpchgen-rs, which emits Int32 for every key except
+-- l_orderkey/o_orderkey, so the parquet and duckdb datasets share one key
+-- layout. It diverges from DuckDB's tpch extension, which makes all keys
+-- BIGINT: a database built here is no longer column-type-interchangeable with
+-- one built by CALL dbgen(). Narrower keys are what let the SF1000 working set
+-- fit -- l_partkey and l_suppkey alone are 48 GB smaller at that scale.
+--
+-- Fixed-width text is VARCHAR, which is how DuckDB stores CHAR anyway.
 
 CREATE TABLE region (
     r_regionkey     INTEGER,
@@ -19,7 +31,7 @@ CREATE TABLE nation (
 );
 
 CREATE TABLE supplier (
-    s_suppkey       BIGINT,
+    s_suppkey       INTEGER,
     s_name          VARCHAR,
     s_address       VARCHAR,
     s_nationkey     INTEGER,
@@ -29,7 +41,7 @@ CREATE TABLE supplier (
 );
 
 CREATE TABLE customer (
-    c_custkey       BIGINT,
+    c_custkey       INTEGER,
     c_name          VARCHAR,
     c_address       VARCHAR,
     c_nationkey     INTEGER,
@@ -40,7 +52,7 @@ CREATE TABLE customer (
 );
 
 CREATE TABLE part (
-    p_partkey       BIGINT,
+    p_partkey       INTEGER,
     p_name          VARCHAR,
     p_mfgr          VARCHAR,
     p_brand         VARCHAR,
@@ -52,8 +64,8 @@ CREATE TABLE part (
 );
 
 CREATE TABLE partsupp (
-    ps_partkey      BIGINT,
-    ps_suppkey      BIGINT,
+    ps_partkey      INTEGER,
+    ps_suppkey      INTEGER,
     ps_availqty     BIGINT,
     ps_supplycost   DECIMAL(15,2),
     ps_comment      VARCHAR
@@ -61,7 +73,7 @@ CREATE TABLE partsupp (
 
 CREATE TABLE orders (
     o_orderkey      BIGINT,
-    o_custkey       BIGINT,
+    o_custkey       INTEGER,
     o_orderstatus   VARCHAR,
     o_totalprice    DECIMAL(15,2),
     o_orderdate     DATE,
@@ -73,8 +85,8 @@ CREATE TABLE orders (
 
 CREATE TABLE lineitem (
     l_orderkey      BIGINT,
-    l_partkey       BIGINT,
-    l_suppkey       BIGINT,
+    l_partkey       INTEGER,
+    l_suppkey       INTEGER,
     l_linenumber    BIGINT,
     l_quantity      DECIMAL(15,2),
     l_extendedprice DECIMAL(15,2),
